Move more constructors to dedicated module
